     \name{data_fs}
     \docType{data}
     \alias{data_fs}
     \title{Some physical characteristics of 92 species}
     \description{
       This data set gives some measurements of ear canals (in millimeters), body mass and agility of 91 known species and 1 fossil record
       compiled by Tim Ryan at Penn State.
     }
     \usage{data_fs}
     \format{A vector containing 92 observations with one missing (FS, a hypothetical fossil record).}
     \source{}
     \references{
     Spoor F, Garland T, Krovitz G, Ryan TM, Silcox MT, and Walker A. (2007)\cr 
     The primate semicircular canal system and locomotion. \emph{ P Natl Acad Sci (USA)}  104:10808-10812.  

     }
     \keyword{datasets}
