     \name{var_fs}
     \docType{data}
     \alias{var_fs}
     \title{Variance-covariance matrix of the 92 species in data_fs}
     \description{
       This data set gives the distance matrix (variance-covariance matrix)
       of the 92 species in data_fs
     }
     \usage{var_fs}
     \format{A matrix of size 92*92.}
     \source{Prepared by Tim Ryan.}
     \references{
       
     }
     \keyword{datasets}
